+++ App.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.11.35312.102
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Apps.XAI", "Apps.X.AI\Apps.XAI.csproj", "{16291B2D-8B02-4B76-9AD9-8E458A466382}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Apps.XAI", "Apps.XAI\Apps.XAI.csproj", "{16291B2D-8B02-4B76-9AD9-8E458A466382}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
