+++ Apps.XAI/Api/XaiRestClient.cs
@@ -1,16 +1,16 @@
-﻿using Apps.X.AI.Models.Response;
+﻿using Apps.XAI.Models.Response;
 using Blackbird.Applications.Sdk.Common.Authentication;
 using Blackbird.Applications.Sdk.Common.Exceptions;
 using Blackbird.Applications.Sdk.Utils.RestSharp;
 using Newtonsoft.Json;
 using RestSharp;
 
-namespace Apps.X.AI.Api
+namespace Apps.XAI.Api
 {
     public class XaiRestClient : BlackBirdRestClient
     {
         public XaiRestClient(IEnumerable<AuthenticationCredentialsProvider> authProviders)
-            : base(new RestClientOptions { ThrowOnAnyError = false, BaseUrl = new Uri("https://api.x.ai/v1") })
+            : base(new RestClientOptions { ThrowOnAnyError = false, BaseUrl = new Uri("https://api.XAI/v1") })
         {
             this.AddDefaultHeader("Authorization", $"Bearer {authProviders.First(x => x.KeyName == "apiKey").Value}");
         }
