+++ Apps.XAI/Apps.XAI.csproj
@@ -6,10 +6,10 @@
         <Nullable>enable</Nullable>
         <Product>xAI [Beta]</Product>
         <Description>xAI allows users to integrate AI-driven completion and chat capabilities into their workflows. Whether you're automating repetitive tasks or generating insights, this app makes advanced AI tools accessible.</Description>
-        <Version>1.0.3</Version>
+        <Version>1.0.4</Version>
         <AssemblyName>Apps.XAI</AssemblyName>
 		<PackageId>Apps.XAI</PackageId>
-		<RootNamespace>Apps.X.AI</RootNamespace>
+		<RootNamespace>Apps.XAI</RootNamespace>
     </PropertyGroup>
 
     <ItemGroup>
